Reorganizes project structure
