+++ main.cpp
@@ -1,5 +1,6 @@
 #include "random_distribution.h"
 #include "electron.h"
+#include "globals.h"
 
 int main(){
     elektron();
